fix(core): fix patient being added after appending extensions
Part of ENG-874
